Refactor integration functionality.
