+++ src/ggrc/integrations/synchronization_jobs/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (C) 2018 Google Inc.
+# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+
+"""Contains integration functionality with issue tracker via cron jobs."""
+
+from ggrc.integrations.synchronization_jobs.assessment_integration import \
+    sync_assessment_statuses  # Noqa
